+++ models/contacts.json
@@ -55,12 +55,8 @@
   },
   {
     "id": "10",
-    "name": "Alec Howard",
+    "name": "Alex Howard",
     "email": "Donec.elementum@scelerisquescelerisquedui.net",
     "phone": "(748) 206-2688"
-  },
-  {
-    "id": "Q25XNgXA9Jp1-_fpFbBZM",
-    "name": "goga"
   }
 ]
