+++ routes/api/contacts.js
@@ -1,77 +1,17 @@
 const express = require("express");
-const contacts = require("../../models/contacts");
-const Joi = require("joi");
+
 const router = express.Router();
-const { RequestError } = require("../../helpers");
-const contactsSchema = Joi.object({
-  name: Joi.string().required(),
-  email: Joi.string().required(),
-  phone: Joi.string().required(),
-});
 
-router.get("/", async (req, res, next) => {
-  try {
-    const result = await contacts.listContacts();
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-});
+const ctrl = require("../../controllers/contacts");
 
-router.get("/:id", async (req, res, next) => {
-  try {
-    const { id } = req.params;
-    const result = await contacts.getContactById(id);
-    if (!result) {
-      throw RequestError(404, "Not found");
-    }
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-});
+router.get("/", ctrl.getAll);
 
-router.post("/", async (req, res, next) => {
-  try {
-    const { error } = contactsSchema.validate(req.body);
-    if (error) {
-      res.status(400).json({ message: "missing required name field" });
-    }
-    const result = await contacts.addContact(req.body);
-    res.status(201).json(result);
-  } catch (error) {
-    next(error);
-  }
-});
+router.get("/:id", ctrl.getById);
 
-router.delete("/:id", async (req, res, next) => {
-  try {
-    const { id } = req.params;
-    const result = await contacts.removeContact(id);
-    if (!result) {
-      throw RequestError(404, "Not found");
-    }
-    res.status(200).json({ message: "contact delited" });
-  } catch (error) {
-    next(error);
-  }
-});
+router.post("/", ctrl.add);
 
-router.put("/:id", async (req, res, next) => {
-  try {
-    const { error } = contactsSchema.validate(req.body);
-    if (error) {
-      throw RequestError(400, error.message);
-    }
-    const { id } = req.params;
-    const result = await contacts.updateContactById(id, req.body);
-    if (!result) {
-      throw RequestError(404, "Not found");
-    }
-    res.status(200).json({ message: "missing fields" });
-  } catch (error) {
-    next(error);
-  }
-});
+router.delete("/:id", ctrl.deleteById);
+
+router.put("/:id", ctrl.updateById);
 
 module.exports = router;
